refactor(onsite_proj): split force/stress wrappers into onsite_proj_force_stress.cpp
Step 3 of file split. Move the four cal_{force,stress}_onsite_{dftu,dspin}
wrappers to onsite_proj_force_stress.cpp. Main onsite_proj.cpp now
contains only the singleton accessor and destructor, with method-level
instantiation for those two. Drop the now-unused dftu_base.h and
<complex> includes from the main file.

Verified by 'make -j 30' in build_max_para_test (built abacus_max_para).
